Added a hashmap with djb2 algorithm, changed dcache so it uses the hashmap. FAT12 is half-finished, procfs needs the fat driver to work for testing it, and devfs could aswell not exist as it only has 2 functions.
Added a fixed-size (256B) boot_cmdline variable in .low.bss that stores variables passed from grub, still need to add a parser for this. This will be the main way to set the root partition as reading that from fstab has a chicken-and-egg problem.

I suggest reading the comments in fs.h and block_device.h.
Please tell me if you find inconsistencies between usages and comments :)

Currently, the allocate_inode and destroy_inode do not have default functions for when their super_ops_t instances are set to NULL.